feat: TopFv 限定だった背景アニメをページ全体に拡張 (#56)
- src/components/SiteBackground/ を新規作成
  - Aurora GLSL シェーダー (Renderer/Program/Mesh) と Particles を集約
  - position: fixed; inset: 0; z-index: -1 で全画面背景化
  - pointer-events: none で前面操作を阻害しない
  - フォールバック背景色 rgba(28, 28, 28, 1) を base に保持
- App.tsx
  - <SiteBackground /> を 1 つだけマウント
  - GlobalStyle の body background を transparent に変更し、
    背景キャンバスが透けて見えるように
- TopFv.tsx
  - VERT/FRAG / Aurora useEffect / Particles JSX を撤去し、
    DecryptedLoader + GlitchText の表示に責務を集約
  - 不要になった ogl / Particles の import 削除
  - TopFvProps と props/colorStops/amplitude/blend/speed も削除
    (背景は SiteBackground が一元管理)
- Aurora.css
  - 未参照になった .aurora-container / .fuzzy-text-container /
    .particles-background ルールを削除
